#168459492 Refactor endpoints - modifies server/src/v1/controllers/userController.js - modifies  server/src/v1/db/createTables.js - modifies  server/src/v1/middleware/schema.js - modifies  server/src/v1/middleware/verify.js - modifies  server/src/v1/tests/sessions.test.js - modifies server/src/v1/tests/user.test.js [finishes #168459492]

+++ server/src/v1/db/createTables.js
@@ -3,7 +3,7 @@ import pool from './configDB';
 const queryText = `
 CREATE TABLE IF NOT EXISTS
 users(
-  id BIGSERIAL,
+  id SERIAL,
   first_name VARCHAR(128) NOT NULL,
   last_name VARCHAR(128) NOT NULL,
   email VARCHAR(128) NOT NULL UNIQUE,
@@ -20,7 +20,7 @@ users(
 CREATE TABLE IF NOT EXISTS
 sessions(
   id SERIAL,
-  mentor_id BIGINT,
+  mentor_id INT,
   mentee_email VARCHAR(128) NOT NULL ,
   questions VARCHAR(200) NOT NULL ,
   status VARCHAR(50) DEFAULT 'Pending',
